requirements.txt
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
fire==0.6.0
fonttools==4.52.4
google-api-core==2.19.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
idna==3.7
imageio==2.34.1
lxml==5.2.2
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pypdfium2==4.30.0
pytesseract==0.3.10
python-docx==1.1.2
rsa==4.9
six==1.16.0
svgwrite==1.4.3
termcolor==2.4.0
typing_extensions==4.12.0
urllib3==2.2.1
